+++ tomcat/src/test/java/org/apache/coyote/http11/request/RequestBodyTest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import static org.assertj.core.api.Assertions.assertThat;
